+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 % erl-uri changelog
 
 # Next Version
+## Features
+- Add URI accessors: `uri:path/1`, `uri:query/1`, `uri:fragment/1`.
 ## Misc
 - Use result tuples for parsing and decoding, instead of signaling errors. For
   example, `uri:parse/1` now returns either `{ok, URI}` or `{error, Reason}`